Gui threaded + other unrelated fixes

+++ cube/swiss/include/gcm.h
@@ -50,6 +50,6 @@ int parse_gcm(file_handle *file, ExecutableFile *filesToPatch);
 void adjust_tgc_fst(char* FST, u32 tgc_base, u32 fileAreaStart, u32 fakeAmount);
 int parse_tgc(file_handle *file, ExecutableFile *filesToPatch, u32 tgc_base, char* tgcname);
 int patch_gcm(file_handle *file, ExecutableFile *filesToPatch, int numToPatch, int multiDol);
-int read_fst(file_handle *file, file_handle** dir, u32 *usedSpace);
+int read_fst(file_handle *file, file_handle** dir, u64 *usedSpace);
 unsigned int getBannerOffset(file_handle *f);
 #endif
